split mechanism into elevator and arm
