+++ src/main/java/frc/robot/BuildConstants.java
@@ -7,12 +7,12 @@ public final class BuildConstants {
   public static final String MAVEN_GROUP = "";
   public static final String MAVEN_NAME = "TB2024";
   public static final String VERSION = "unspecified";
-  public static final int GIT_REVISION = 61;
-  public static final String GIT_SHA = "8a315593a5c629aad569144d4224dacb557eaadb";
-  public static final String GIT_DATE = "2024-02-03 23:04:21 EST";
+  public static final int GIT_REVISION = 62;
+  public static final String GIT_SHA = "1a04ae7005c85d172d1e41a745a43ae3be64d2a4";
+  public static final String GIT_DATE = "2024-02-06 18:19:10 EST";
   public static final String GIT_BRANCH = "Mechanism-Sim";
-  public static final String BUILD_DATE = "2024-02-06 18:13:52 EST";
-  public static final long BUILD_UNIX_TIME = 1707261232727L;
+  public static final String BUILD_DATE = "2024-02-06 19:22:28 EST";
+  public static final long BUILD_UNIX_TIME = 1707265348206L;
   public static final int DIRTY = 1;
 
   private BuildConstants(){}
